requirements.txt
Django==1.7.1
Pillow==2.8.1
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.4.0
django-crispy-forms==1.4.0
django-s3direct==0.3.9
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.1.1
mechanize==0.2.5
psycopg2==2.5.4
pystache==0.5.4
requests==2.6.0
six==1.9.0
static==1.1.1
static3==0.5.1
wsgiref==0.1.2
